modify grub attributes and disable grub recipe by default
if the vm is in a private cloud (e.g. VMware) and the vm has
cloud-init installed (e.g. ubuntu cloud images) but there is no
user-data datasource then on boot cloud-init will take a long time
searching for a datasource that doesn't exist

setting grub attributes tells cloud-init to skip the datasource search

i disabled the grub recipe rather than assume that users of this
chef-repo will not have a datasource